# FoodBook/FoodBook/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FoodBookAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Difficulties": {
      "get": {
        "tags": [
          "Difficulties"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Difficulty"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Difficulty"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Difficulty"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Difficulties"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Difficulty"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Difficulty"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Difficulty"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Difficulty"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Difficulty"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Difficulty"
                }
              }
            }
          }
        }
      }
    },
    "/api/Difficulties/{id}": {
      "get": {
        "tags": [
          "Difficulties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Difficulty"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Difficulty"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Difficulty"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Difficulties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Difficulty"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Difficulty"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Difficulty"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Difficulties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DishTypes": {
      "get": {
        "tags": [
          "DishTypes"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishTypeForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishTypeForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishTypeForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DishTypes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishTypeForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DishTypeForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DishTypeForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DishTypeForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishTypeForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishTypeForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/DishTypes/{id}": {
      "get": {
        "tags": [
          "DishTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DishTypeForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishTypeForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishTypeForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "DishTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishTypeForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DishTypeForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DishTypeForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "DishTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Difficulty": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "DishTypeForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "imageURL": {
            "type": "string",
            "nullable": true
          },
          "idImage": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "imageTitle": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}